requirements.txt
# ===== Core runtime =====
python-dotenv==1.1.1
tqdm==4.67.1
rich==14.1.0
pandas==2.3.2
numpy==1.26.4
pyarrow==15.0.2

# ===== API & Monitoring =====
fastapi==0.114.2
uvicorn[standard]==0.30.6
prometheus-client==0.20.0
pydantic==2.9.2

# ===== Vector DB =====
qdrant-client==1.12.2

# ===== Embeddings / Retrieval / Reranking =====
sentence-transformers==3.0.1
torch==2.3.1
scikit-learn==1.5.2
transformers==4.44.2
accelerate==0.33.0

# ===== LangChain / LangGraph =====
langchain==0.3.8
langgraph==0.2.35
langchain-openai==0.3.3

# ===== OpenAI SDK (Responses API) =====
openai==1.107.3

# ===== RAG evaluation =====
ragas==0.3.4
datasets==4.0.0
evaluate==0.4.2

# ===== Monitoring =====
prometheus-client==0.20.0
prometheus-fastapi-instrumentator==6.1.0

# ===== Document parsing (txt/pdf/docx) =====
pymupdf==1.24.9
python-docx==1.1.2
PyYAML==6.0.2

# ===== MLflow (tracking + artifacts) =====
mlflow==2.14.1

# ===== Graph visualization & NER =====
networkx==3.2.1
graphviz==0.21
pydot==4.0.1
matplotlib==3.9.4
natasha==1.6.0
razdel==0.5.0     